Jury rigged Debugger library to build
Added an integration test to test that the library links correctly.
Couple source code hacks to get around warnings that weren't previously
flagged in the previous project.
Actually builds a library instead of the dummy executable.

+++ src/source/DebuggerInterpreter.h
@@ -33,7 +33,7 @@ private:
 
     bool m_listNext = false;
     unsigned int m_listsize = 10;
-    unsigned int m_listAddress = 0;
+    size_t m_listAddress = 0;
     std::string m_commandResponse;
     Debugger* m_debugger;
 };
